Fix pull-request-target changelog requirements (#7)
I implemented a workflow to check whether a pull request had a news
fragment but I did it with `pull_request_target`, which doesn't have
access to the code from the PR. It's really just for doing surface level
checks of PRs that you can always trust to run. Make it a normal
`pull_request` trigger.